Move page styles next to source code
