+++ TF_Settings_Web/src/Components/Pages/Camera/Camera.scss
