+++ TF_Settings_Web/src/Components/Pages/Camera/ManualSetup/ManualSetup.scss
