+++ TF_Settings_Web/src/Components/Pages/Camera/QuickSetup/Calibration/CalibrationCompleteScreen.tsx
@@ -1,4 +1,4 @@
-import 'Styles/Camera/Calibrate.scss';
+import './Calibration.scss';
 import cssVariables from 'Styles/variables.module.scss';
 
 import React, { CSSProperties, useEffect, useRef, useState } from 'react';
