+++ TF_Settings_Web/src/Components/Pages/Camera/QuickSetup/Calibration/CalibrationManager.tsx
@@ -1,4 +1,4 @@
-import 'Styles/Camera/Calibrate.scss';
+import './Calibration.scss';
 
 import React, { useEffect } from 'react';
 import { Route, Routes, useNavigate } from 'react-router-dom';
