+++ TF_Settings_Web/src/Components/Pages/Camera/QuickSetup/QuickSetupManager.tsx
@@ -1,4 +1,4 @@
-import 'Styles/Camera/Camera.scss';
+import '@/Components/Pages/Camera/Camera.scss';
 
 import React from 'react';
 import { Route, Routes } from 'react-router-dom';
