+++ TF_Settings_Web/src/Components/Pages/Interactions/InteractionsScreen.tsx
@@ -1,4 +1,4 @@
-import 'Styles/Interactions.scss';
+import './Interactions.scss';
 
 import { Component } from 'react';
