+++ TF_Settings_Web/src/Components/StatusIndicator/StatusIndicator.scss
