+++ TF_Settings_Web/src/Components/StatusIndicator/StatusIndicator.tsx
@@ -1,4 +1,4 @@
-import 'Styles/StatusIndicator.scss';
+import './StatusIndicator.scss';
 
 import React from 'react';
