+++ TF_Settings_Web/src/Components/VersionIndicator/VersionIndicator.scss
