+++ TF_Settings_Web/src/Components/VersionIndicator/VersionIndicator.tsx
@@ -1,4 +1,4 @@
-import 'Styles/VersionIndicator.scss';
+import './VersionIndicator.scss';
 
 import React from 'react';
